+++ front/app/components/admin/ProjectableHeaderBgUploader/ProjectableHeaderBgUploader.test.tsx
@@ -2,8 +2,6 @@ import React from 'react';
 import { screen, render, waitFor, fireEvent } from 'utils/testUtils/rtl';
 import HeaderBgUploader from './';
 
-jest.mock('utils/cl-intl');
-
 const file = new File(['file'], 'file.jpeg', {
   type: 'image/jpeg',
 });