+++ front/app/containers/Admin/pagesAndMenu/containers/GenericHeroBannerForm/BannerImageFields/messages.ts
@@ -1,0 +1,18 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  imageSupportPageURL: {
+    id: 'pagesAndMenu.GenericHeroBannerForm.BannerImageFields.imageSupportPageURL',
+    defaultMessage:
+      'https://support.citizenlab.co/en/articles/1346397-what-are-the-recommended-dimensions-and-sizes-of-the-platform-images',
+  },
+  imageTooltip: {
+    id: 'pagesAndMenu.GenericHeroBannerForm.BannerImageFields.imageTooltip',
+    defaultMessage:
+      'For more information on recommended image resolutions, {supportPageLink}.',
+  },
+  imageSupportPageText: {
+    id: 'pagesAndMenu.GenericHeroBannerForm.BannerImageFields.imageSupportPageText',
+    defaultMessage: 'visit our support center',
+  },
+});