+++ front/app/containers/MobileNavbar/MobileNavbar.test.tsx
@@ -2,12 +2,9 @@ import React from 'react';
 import { render, screen } from 'utils/testUtils/rtl';
 import MobileNavbar from '.';
 
-jest.mock('services/locale');
-jest.mock('services/appConfiguration');
 jest.mock('hooks/useNavbarItems');
 jest.mock('utils/analytics');
 jest.mock('utils/cl-router/withRouter');
-jest.mock('utils/cl-router/Link');
 
 describe('<MobileNavbar />', () => {
   it('renders', () => {