+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 2023-02-10
+
+### Changed
+
+- Tooltips for image uploaders in project and folder forms.
+- [CL-737] Show login screen after successful password reset
+
+## 2023-02-09
+
+### Added
+
+- Report builder templates now automatically detect ideation and native survey projects and add different widgets based on that
+- [CL-2739] Added basic posthog integration
+
+### Changed
+
+- [CL-2673] In-platform survey UI improvements
+
 ## 2023-02-03
 
 ### Added